# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Div test exercise API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/requests": {
            "get": {
                "tags": [
                    "User Requests"
                ],
                "summary": "Get user requests",
                "description": "Get all records of user requests",
                "operationId": "getUserRequests",
                "parameters": [
                    {
                        "name": "per-page",
                        "in": "query",
                        "description": "Records per pagee",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "filters[status]",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "example": "Active"
                    },
                    {
                        "name": "filters[created-at][from]",
                        "in": "query",
                        "description": "Filter by date created",
                        "required": false,
                        "example": "22.07.2023"
                    },
                    {
                        "name": "filters[created-at][to]",
                        "in": "query",
                        "description": "Filter by date created",
                        "required": false,
                        "example": "23.07.2023"
                    },
                    {
                        "name": "user-name",
                        "in": "query",
                        "description": "Filter by user name",
                        "required": false,
                        "example": "Test user"
                    },
                    {
                        "name": "user-email",
                        "in": "query",
                        "description": "Filter by user email",
                        "required": false,
                        "example": "test@gmail.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UserRequest"
                                                    }
                                                },
                                                "links": {
                                                    "properties": {
                                                        "first": {
                                                            "type": "string",
                                                            "example": "http://localhost/api/v1/requests?page=1"
                                                        },
                                                        "last": {
                                                            "type": "string",
                                                            "example": "http://localhost/api/v1/requests?page=10"
                                                        },
                                                        "prev": {
                                                            "type": "string",
                                                            "example": "http://localhost/api/v1/requests?page=2"
                                                        },
                                                        "next": {
                                                            "type": "string",
                                                            "example": "http://localhost/api/v1/requests?page=4"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Requests"
                ],
                "summary": "Send request",
                "description": "Send request",
                "operationId": "sendUserRequest",
                "requestBody": {
                    "description": "User request info fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test user"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@gmail.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Test message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "request_id": {
                                                    "type": "integer",
                                                    "example": "5"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{user_request_id}": {
            "get": {
                "tags": [
                    "User Requests"
                ],
                "summary": "Get user request by id",
                "description": "Get one record of user request by id",
                "operationId": "getUserRequest",
                "parameters": [
                    {
                        "name": "user_request_id",
                        "in": "path",
                        "description": "User request id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserRequest"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Requests"
                ],
                "summary": "Resolve user request",
                "description": "Resolve user request by manager",
                "operationId": "resolveUserRequest",
                "parameters": [
                    {
                        "name": "user_request_id",
                        "in": "path",
                        "description": "User request id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Manager comment, that sended to mail user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Your propblem has been solved!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": "[]"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Requests"
                ],
                "summary": "Delete user request",
                "description": "Delete user request by manager",
                "operationId": "deleteUserRequest",
                "parameters": [
                    {
                        "name": "user_request_id",
                        "in": "path",
                        "description": "User request id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": "[]"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "readOnly": "true",
                        "example": "Test user"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "role_id": {
                        "description": "User role id",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "UserRequest": {
                "required": [
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "readOnly": "true",
                        "example": "Test user"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "status": {
                        "description": "User request status",
                        "type": "string",
                        "readOnly": "true",
                        "example": "Active"
                    },
                    "comment": {
                        "description": "Comment of manager. Not null only if status = Resolved",
                        "type": "string",
                        "example": "Test commnet of manager"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserRequest"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}